Split repository.ts: extract the scoped session/digest/gap domains (12/N)
Twelfth slice of the repository.ts carve-up (audit L14), and the first to move
🔒 conversation-scoped admin reads — so this one carries extra verification.

- repository/sessions.ts (166) — session continuity plus the recall reads
  (searchMemory / userMessages) behind memory context.
- repository/questionDigest.ts (59) — question clustering for the admin digest.
- repository/knowledgeGaps.ts (276) — below-floor knowledge_search misses (#208),
  their clustering and escalation state.

Also promotes two cross-cutting internals into repository/shared.ts, both used
by three or more of the remaining domains: `cosineSim` and
`QUESTION_CLUSTER_SIMILARITY_THRESHOLD` (0.85, the "is this the same question"
duplicate floor — deliberately distinct from the 0.35 relevance floor already
there). cosineSim is identical modulo the added `export`.

🔒 SCOPING VERIFICATION. These modules own reads that must stay narrowed to the
conversations an admin is actually in — the invariant CLAUDE.md states as "admin
data access is scoped in SQL to conversations the admin is in", null meaning
super_admin/unrestricted. I checked mechanically rather than by eye:
- every `conversationIds` parameter form in the moved code
  (`readonly string[] | null`, `?: readonly string[]`) appears verbatim in
  origin/main;
- the `conversation_id = ANY($n)` predicate appears verbatim;
- no scoped function lost its parameter, and no predicate was reformatted.
Each module's header now states the scoping contract so the next reader of a
1-domain file sees it without reading repository.ts.

18/19 moved symbols byte-identical (the 19th is cosineSim, identical modulo
`export`). repository.ts is now 2,089 lines, down from 7,137 — 71% reduction.
