+++ src/storage/repository.ts
@@ -34,7 +34,11 @@ export * from './repository/shared.js';
 export * from './repository/devTeamWatches.js';
 // `export *` does not bind names into this module's own scope, so anything
 // still living here that uses an extracted symbol must import it explicitly.
-import { invalidateDigestsForInteractions } from './repository/shared.js';
+import {
+  cosineSim,
+  invalidateDigestsForInteractions,
+  QUESTION_CLUSTER_SIMILARITY_THRESHOLD,
+} from './repository/shared.js';
 import { AUTO_ENROLL_ACTOR } from './repository/members.js';
 import { sumShortcutHits } from './repository/shortcutHits.js';
 export * from './repository/accessRequests.js';
@@ -53,6 +57,9 @@ export * from './repository/knowledge.js';
 export * from './repository/knowledgeCandidates.js';
 export * from './repository/suggestions.js';
 export * from './repository/budgetsPrivacy.js';
+export * from './repository/sessions.js';
+export * from './repository/questionDigest.js';
+export * from './repository/knowledgeGaps.js';
 
 export interface InteractionInput {
   platform: Platform;
@@ -448,172 +455,6 @@ export async function getInteractionContentByMessageId(
   return rows[0]?.content ?? null;
 }
 
-// --- Sessions --------------------------------------------------------------
-
-export interface StoredSession {
-  sessionId: string;
-  turnCount: number;
-  updatedAt: Date;
-}
-
-export async function getClaudeSession(
-  platform: Platform,
-  conversationId: string,
-): Promise<StoredSession | null> {
-  let rows: Array<{ claude_session_id: string | null; turn_count: unknown; updated_at: Date }>;
-  try {
-    ({ rows } = await pool.query(
-      `SELECT claude_session_id, turn_count, updated_at
-       FROM sessions WHERE platform = $1 AND conversation_id = $2`,
-      [platform, conversationId],
-    ));
-  } catch (err) {
-    // Degrade to "no stored session" so the turn starts fresh instead of
-    // dying — runAgentTurn already treats null as start-fresh (issue #52).
-    logger.warn({ err }, 'Session lookup failed; starting a fresh session');
-    return null;
-  }
-  const row = rows[0];
-  if (!row?.claude_session_id) return null;
-  return {
-    sessionId: row.claude_session_id,
-    turnCount: Number(row.turn_count ?? 0),
-    updatedAt: row.updated_at,
-  };
-}
-
-/** Upsert the session id; the turn counter increments on resume, resets on a new session. */
-export async function setClaudeSessionId(
-  platform: Platform,
-  conversationId: string,
-  sessionId: string,
-): Promise<void> {
-  await pool.query(
-    `INSERT INTO sessions (platform, conversation_id, claude_session_id, turn_count)
-     VALUES ($1, $2, $3, 1)
-     ON CONFLICT (platform, conversation_id)
-     DO UPDATE SET
-       turn_count = CASE
-         WHEN sessions.claude_session_id = EXCLUDED.claude_session_id
-         THEN sessions.turn_count + 1 ELSE 1 END,
-       claude_session_id = EXCLUDED.claude_session_id`,
-    [platform, conversationId, sessionId],
-  );
-}
-
-/** Drop a stored session id (e.g. after a failed resume) so the next turn starts fresh. */
-export async function clearClaudeSessionId(platform: Platform, conversationId: string): Promise<void> {
-  await pool.query(
-    `UPDATE sessions SET claude_session_id = NULL, updated_at = now()
-      WHERE platform = $1 AND conversation_id = $2`,
-    [platform, conversationId],
-  );
-}
-
-/**
- * Reset the stored Claude session for every conversation the given user is
- * active in on `platform`, so a role change (grant_admin/revoke_admin) takes
- * effect on their very next message instead of being shadowed by the old-role
- * framing still in a live session's history until it rolls over
- * (SESSION_MAX_TURNS/AGE). Without this, a freshly-promoted admin keeps getting
- * refused, and — more importantly — a freshly-*revoked* admin's session could
- * keep treating them as admin for up to a full session's worth of turns.
- *
- * Non-destructive: only clears session *continuity* (nulls `claude_session_id`,
- * same primitive as `clearClaudeSessionId`); stored interactions/memory are
- * untouched and the next turn rebuilds context from them. Scoped to
- * conversations the user has actually participated in — in a group that means
- * the group's shared thread resets, which is the same fresh-start that happens
- * on normal rollover. Returns the number of sessions cleared.
- */
-export async function clearUserSessions(platform: Platform, userId: string): Promise<number> {
-  const { rowCount } = await pool.query(
-    `UPDATE sessions
-        SET claude_session_id = NULL, updated_at = now()
-      WHERE platform = $1
-        AND claude_session_id IS NOT NULL
-        AND conversation_id IN (
-          SELECT DISTINCT conversation_id FROM interactions
-           WHERE platform = $1 AND user_id = $2
-        )`,
-    [platform, userId],
-  );
-  return rowCount ?? 0;
-}
-
-/**
- * True if the bot has previously seen this conversation on this platform.
- * Used to stop privileged tools from targeting arbitrary ids (e.g. messaging
- * any phone number on WhatsApp).
- */
-export async function isKnownConversation(platform: Platform, conversationId: string): Promise<boolean> {
-  const { rows } = await pool.query(
-    `SELECT 1 FROM interactions WHERE platform = $1 AND conversation_id = $2 LIMIT 1`,
-    [platform, conversationId],
-  );
-  return rows.length > 0;
-}
-
-/**
- * Recent messages by a user, optionally restricted to a conversation set
- * (admin scoping: an admin only sees history from conversations they share).
- */
-export async function userMessages(
-  platform: Platform,
-  userId: string,
-  limit = 20,
-  conversationIds?: readonly string[],
-): Promise<Array<{ conversationId: string; direction: string; content: string; createdAt: Date }>> {
-  const params: unknown[] = [platform, userId];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    scope = `AND conversation_id = ANY($${params.length})`;
-  }
-  params.push(limit);
-  const { rows } = await pool.query(
-    `SELECT conversation_id, direction, content, created_at
-       FROM interactions
-      WHERE platform = $1 AND user_id = $2 ${scope}
-      ORDER BY created_at DESC
-      LIMIT $${params.length}`,
-    params,
-  );
-  return rows.map((r) => ({
-    conversationId: r.conversation_id,
-    direction: r.direction,
-    content: r.content,
-    createdAt: r.created_at,
-  }));
-}
-
-/** True if the bot has previously seen this user on this platform. */
-export async function isKnownUser(platform: Platform, userId: string): Promise<boolean> {
-  const { rows } = await pool.query(
-    `SELECT 1 FROM interactions WHERE platform = $1 AND user_id = $2 LIMIT 1`,
-    [platform, userId],
-  );
-  return rows.length > 0;
-}
-
-/**
- * True if the bot has stored this exact message id within this conversation
- * (issue #231: `react_to_message`'s target validation — same "the bot must
- * have actually seen it" discipline as `isKnownUser`/`isKnownConversation`,
- * scoped to one conversation since a member may only react within their own).
- */
-export async function isKnownMessage(
-  platform: Platform,
-  conversationId: string,
-  messageId: string,
-): Promise<boolean> {
-  const { rows } = await pool.query(
-    `SELECT 1 FROM interactions WHERE platform = $1 AND conversation_id = $2 AND message_id = $3 LIMIT 1`,
-    [platform, conversationId, messageId],
-  );
-  return rows.length > 0;
-}
-
 // --- Super-admin views ---------------------------------------------------------
 
 export async function recentAuditEntries(limit = 20): Promise<
@@ -957,351 +798,6 @@ export async function sumBackgroundJobCosts(
   );
   const byJob = rows.map((r) => ({ job: r.job as string, costUsd: Number(r.cost) }));
   return { total: byJob.reduce((sum, r) => sum + r.costUsd, 0), byJob };
-}
-
-// --- Question digest ---------------------------------------------------------
-
-export interface QuestionCluster {
-  representative: string;
-  count: number;
-}
-
-const QUESTION_CLUSTER_SIMILARITY_THRESHOLD = 0.85;
-
-/** Dot product of two embed()-produced (L2-normalized) vectors equals cosine similarity. */
-function cosineSim(a: number[], b: number[]): number {
-  let dot = 0;
-  const n = Math.min(a.length, b.length);
-  for (let i = 0; i < n; i++) dot += a[i] * b[i];
-  return dot;
-}
-
-/**
- * Greedily cluster recently-addressed inbound messages by embedding
- * similarity to surface recurring, un-curated questions — a signal for what
- * should become a `knowledge` entry. Clustering runs in application code over
- * an already time-bounded, conversation-scoped result set (no SQL self-join;
- * see #21 for why that's the right tradeoff at this scale).
- */
-export async function recentQuestionClusters(
-  conversationIds: readonly string[] | null,
-  days = 7,
-  limit = 10,
-): Promise<QuestionCluster[]> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 30);
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 10, 1), 50);
-
-  const params: unknown[] = [`${clampedDays} days`];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    scope = `AND conversation_id = ANY($${params.length})`;
-  }
-
-  const { rows } = await pool.query(
-    `SELECT content, embedding
-       FROM interactions
-      WHERE addressed_to_bot = true AND direction = 'inbound'
-        AND embedding IS NOT NULL
-        AND created_at > now() - $1::interval
-        ${scope}
-      ORDER BY created_at ASC`,
-    params,
-  );
-
-  const clusters: Array<{ representative: string; embedding: number[]; count: number }> = [];
-  for (const row of rows) {
-    const vec = row.embedding as number[] | null;
-    if (!vec) continue;
-    const match = clusters.find((c) => cosineSim(c.embedding, vec) >= QUESTION_CLUSTER_SIMILARITY_THRESHOLD);
-    if (match) {
-      match.count += 1;
-    } else {
-      clusters.push({ representative: row.content, embedding: vec, count: 1 });
-    }
-  }
-
-  return clusters
-    .filter((c) => c.count >= 2)
-    .sort((a, b) => b.count - a.count)
-    .slice(0, clampedLimit)
-    .map((c) => ({ representative: c.representative, count: c.count }));
-}
-
-// --- Knowledge gaps (below-floor knowledge_search misses, issue #208) -------
-
-/** Per-user cap on new gap rows within a rolling 24h window — same anti-flood shape as RATE_ANSWER_DAILY_LIMIT. */
-export const KNOWLEDGE_GAP_DAILY_LIMIT = 20;
-export const KNOWLEDGE_GAP_QUERY_MAX_CHARS = 500;
-
-/**
- * Record one `knowledge_search` call that came back with hits but none
- * cleared `KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD` — the caller (the
- * `knowledge_search` tool handler) must only invoke this when
- * `hits.length > 0 && relevantIds.length === 0`, never on a plain empty
- * result set, so a `searchKnowledge` embed() failure (which also returns
- * `[]`) can't masquerade as a genuine miss. `query` is the model's
- * reformulated search string, not necessarily the member's verbatim
- * message — callers/docs must describe entries as "searches with no
- * confident answer", not "member questions".
- *
- * Enforces the same DB-backed rolling-24h cap per `(platform, user_id)` as
- * `createAnswerFeedback`/`createSuggestion` (COUNT(*) inside the insert,
- * never an in-memory counter) so a chatty or adversarial member can't flood
- * `list_knowledge_gaps` with junk. Fire-and-forget from the tool handler —
- * callers must swallow failures themselves (never block or delay the reply).
- */
-export async function recordKnowledgeGap(
-  platform: Platform,
-  conversationId: string,
-  userId: string,
-  query: string,
-): Promise<{ id: number } | 'rate_limited'> {
-  let embedding: number[] | null = null;
-  try {
-    embedding = await embed(query);
-  } catch (err) {
-    logger.warn({ err }, 'Embedding failed for knowledge gap');
-  }
-
-  const { rows } = await pool.query(
-    `WITH recent AS (
-       SELECT count(*) AS n FROM knowledge_gaps
-        WHERE platform = $1 AND user_id = $2
-          AND created_at > now() - interval '24 hours'
-     )
-     INSERT INTO knowledge_gaps (platform, conversation_id, user_id, query_text, embedding)
-     SELECT $1, $3, $2, $4, $5
-      WHERE (SELECT n FROM recent) < $6
-     RETURNING id`,
-    [
-      platform,
-      userId,
-      conversationId,
-      query.slice(0, KNOWLEDGE_GAP_QUERY_MAX_CHARS),
-      embedding ? pgvector.toSql(embedding) : null,
-      KNOWLEDGE_GAP_DAILY_LIMIT,
-    ],
-  );
-  return rows[0] ? { id: Number(rows[0].id) } : 'rate_limited';
-}
-
-/**
- * Record a CONFIRMED escalation (issue #479's escalation-confirmation
- * intercept) into `knowledge_gaps` with `escalated = true` — the strongest
- * curation-priority signal available: a member asked a human directly,
- * rather than a passive below-floor `knowledge_search` miss (issue #514).
- * Deliberately an unconditional insert, NOT gated by `KNOWLEDGE_GAP_DAILY_LIMIT`
- * — that per-user cap exists to bound passive per-message noise, and reusing
- * it here would risk silently dropping the highest-value data point. The
- * caller (router.ts) only ever invokes this inside the
- * `reserveEscalationSlot` success branch, so volume is already independently
- * bounded by the guild-wide `ESCALATION_RATE_LIMIT_PER_HOUR`. Fire-and-forget
- * from the router — callers must swallow failures themselves (never block or
- * delay the confirmation reply), matching the sibling `notifyAdminsFn` call.
- */
-export async function recordEscalatedKnowledgeGap(
-  platform: Platform,
-  conversationId: string,
-  userId: string,
-  query: string,
-): Promise<{ id: number }> {
-  let embedding: number[] | null = null;
-  try {
-    embedding = await embed(query);
-  } catch (err) {
-    logger.warn({ err }, 'Embedding failed for escalated knowledge gap');
-  }
-
-  const { rows } = await pool.query(
-    `INSERT INTO knowledge_gaps (platform, conversation_id, user_id, query_text, embedding, escalated)
-     VALUES ($1, $2, $3, $4, $5, true)
-     RETURNING id`,
-    [
-      platform,
-      conversationId,
-      userId,
-      query.slice(0, KNOWLEDGE_GAP_QUERY_MAX_CHARS),
-      embedding ? pgvector.toSql(embedding) : null,
-    ],
-  );
-  return { id: Number(rows[0].id) };
-}
-
-export interface KnowledgeGapCluster {
-  representative: string;
-  count: number;
-}
-
-/**
- * Greedily cluster recent knowledge-search misses by embedding similarity —
- * the `list_knowledge_gaps` signal, mirroring `recentQuestionClusters` exactly
- * (same clustering code, same `QUESTION_CLUSTER_SIMILARITY_THRESHOLD`,
- * same conversation-scoping convention) but sourced from `knowledge_gaps`
- * instead of `interactions`. Excludes `resolved_at IS NOT NULL` rows (issue
- * #422) — a gap `save_knowledge`/`update_knowledge` already resolved
- * disappears immediately, not only once `created_at` ages past `days`.
- */
-export async function recentKnowledgeGapClusters(
-  conversationIds: readonly string[] | null,
-  days = 7,
-  limit = 10,
-): Promise<KnowledgeGapCluster[]> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 30);
-  const clampedLimit = Math.min(Math.max(Math.trunc(limit) || 10, 1), 50);
-
-  const params: unknown[] = [`${clampedDays} days`];
-  let scope = '';
-  if (conversationIds) {
-    params.push([...conversationIds]);
-    scope = `AND conversation_id = ANY($${params.length})`;
-  }
-
-  const { rows } = await pool.query(
-    `SELECT query_text, embedding
-       FROM knowledge_gaps
-      WHERE embedding IS NOT NULL
-        AND resolved_at IS NULL
-        AND created_at > now() - $1::interval
-        ${scope}
-      ORDER BY created_at ASC`,
-    params,
-  );
-
-  const clusters: Array<{ representative: string; embedding: number[]; count: number }> = [];
-  for (const row of rows) {
-    const vec = row.embedding as number[] | null;
-    if (!vec) continue;
-    const match = clusters.find((c) => cosineSim(c.embedding, vec) >= QUESTION_CLUSTER_SIMILARITY_THRESHOLD);
-    if (match) {
-      match.count += 1;
-    } else {
-      clusters.push({ representative: row.query_text, embedding: vec, count: 1 });
-    }
-  }
-
-  return clusters
-    .filter((c) => c.count >= 2)
-    .sort((a, b) => b.count - a.count)
-    .slice(0, clampedLimit)
-    .map((c) => ({ representative: c.representative, count: c.count }));
-}
-
-export interface CrossedKnowledgeGapCluster extends KnowledgeGapCluster {
-  /** ids of every row in the crossed cluster, for `markKnowledgeGapsAlerted`. */
-  rowIds: number[];
-}
-
-/**
- * Real-time counterpart to `recentKnowledgeGapClusters` above (issue #650):
- * identical greedy embedding-similarity clustering, scoped to `newGapId`'s
- * own conversation, but restricted to `alerted_at IS NULL` rows so a cluster
- * already promoted to an alert (its rows stamped by `markKnowledgeGapsAlerted`)
- * can't re-trigger on a later gap — only fresh, not-yet-alerted rows count
- * toward a new crossing. Called synchronously right after `recordKnowledgeGap`'s
- * insert, scoped to that one gap's own conversation (never guild-wide),
- * matching `recentKnowledgeGapClusters`'s own scoping convention.
- *
- * Returns the crossed cluster's `representative`/`count` (same shape as
- * `recentKnowledgeGapClusters`) plus every member row's `id`, so the caller
- * can stamp them all alerted in one `markKnowledgeGapsAlerted` call.
- * `recentKnowledgeGapClusters` itself is left untouched — still backs
- * `list_knowledge_gaps`, never returns row ids, since exposing ids there
- * would be unused surface for that read-only listing path.
- *
- * Returns null when `newGapId`'s own cluster (identified by embedding
- * similarity) hasn't yet reached `threshold` unalerted rows, or when the new
- * row has no embedding (an `embed()` failure at insert time — no signal to
- * cluster on) or isn't found among the unresolved/unalerted rows (already
- * resolved or alerted by a race with another turn).
- */
-export async function findCrossedKnowledgeGapCluster(
-  conversationId: string,
-  newGapId: number,
-  threshold: number,
-  days = 7,
-): Promise<CrossedKnowledgeGapCluster | null> {
-  const clampedDays = Math.min(Math.max(Math.trunc(days) || 7, 1), 30);
-
-  const { rows } = await pool.query(
-    `SELECT id, query_text, embedding
-       FROM knowledge_gaps
-      WHERE conversation_id = $1
-        AND resolved_at IS NULL
-        AND alerted_at IS NULL
-        AND embedding IS NOT NULL
-        AND created_at > now() - $2::interval
-      ORDER BY created_at ASC`,
-    [conversationId, `${clampedDays} days`],
-  );
-
-  const clusters: Array<{ representative: string; embedding: number[]; count: number; ids: number[] }> = [];
-  for (const row of rows) {
-    const vec = row.embedding as number[] | null;
-    if (!vec) continue;
-    const match = clusters.find((c) => cosineSim(c.embedding, vec) >= QUESTION_CLUSTER_SIMILARITY_THRESHOLD);
-    if (match) {
-      match.count += 1;
-      match.ids.push(Number(row.id));
-    } else {
-      clusters.push({ representative: row.query_text, embedding: vec, count: 1, ids: [Number(row.id)] });
-    }
-  }
-
-  const crossed = clusters.find((c) => c.ids.includes(newGapId) && c.count >= threshold);
-  if (!crossed) return null;
-  return { representative: crossed.representative, count: crossed.count, rowIds: crossed.ids };
-}
-
-/**
- * Stamps `alerted_at = now()` on every row of a just-crossed cluster (issue
- * #650), returned by `findCrossedKnowledgeGapCluster` — single-shot per
- * cluster: that function's `alerted_at IS NULL` filter means none of these
- * rows can contribute to a future crossing again. Called unconditionally once
- * the caller has reserved a real-time-alert rate-limit slot, regardless of
- * whether the subsequent admin DM itself succeeds — same "the slot is
- * consumed the moment we decide to alert" precedent as
- * `reserveEscalationSlot`/`reserveAccessRequestAlertSlot`.
- */
-export async function markKnowledgeGapsAlerted(ids: readonly number[]): Promise<void> {
-  if (ids.length === 0) return;
-  await pool.query(`UPDATE knowledge_gaps SET alerted_at = now() WHERE id = ANY($1)`, [[...ids]]);
-}
-
-/**
- * Real-time stale-knowledge admin nudge (issue #701): atomically checks the
- * re-arm gate — `stale_alerted_at IS NULL OR stale_alerted_at < updated_at` —
- * and, if it passes, stamps `stale_alerted_at = now()` in the SAME statement,
- * so two concurrent calls for the same row (e.g. two knowledge_search hits in
- * one turn) can't both pass the gate. Returns the row's `title`/`content`/
- * `updatedAt` (enough for the caller to build the DM body via
- * `formatRelativeAge`/`truncateForEcho`) only when the stamp happened; `null`
- * when the row was already alerted since its last edit — an admin edit via
- * `update_knowledge`/`accept_knowledge_candidate` bumps `updated_at` (the
- * `knowledge_set_updated_at` trigger) and so re-arms the gate automatically,
- * no separate reset needed.
- *
- * Called unconditionally by the caller whenever a served hit is stale,
- * REGARDLESS of whether the guild-wide rate limit has a free slot — the
- * opposite of `markKnowledgeGapsAlerted`'s "only stamp once the alert is
- * actually reserved" precedent. This must always stamp so a rate-limited
- * entry doesn't retry-storm on every subsequent serve for as long as it stays
- * stale (acceptance criterion 5c); the rate limit only ever gates the
- * `notifyAdmins` call itself, never this stamp.
- */
-export async function markStaleKnowledgeAlerted(
-  id: number,
-): Promise<{ title: string | null; content: string; updatedAt: Date } | null> {
-  const { rows } = await pool.query(
-    `UPDATE knowledge
-        SET stale_alerted_at = now()
-      WHERE id = $1
-        AND (stale_alerted_at IS NULL OR stale_alerted_at < updated_at)
-      RETURNING title, content, updated_at`,
-    [id],
-  );
-  if (rows.length === 0) return null;
-  return { title: rows[0].title, content: rows[0].content, updatedAt: rows[0].updated_at };
 }
 
 // --- Member-facing weekly digest freshness guard (issue #645) --------------
